Fix random crashes observed during VTS execution
mNgraphNodes was a static member of OperationsBase class.
This was because all operations extend the OperationBase class,
and a separate copy for each operation is not required.
But, when running the entire VTS suite for neuralnetworks,
multiple instances are created from different threads, and
executed parallelly. This causes corruption of the mNgraphNodes
instance, resulting in random crashes.
Hence, making mNgraphNodes an non-static member of the class, so
that separate instances are maintained by each thread.

Signed-off-by: Anoob Anto K <anoob.anto.kodankandath@intel.com>