remove mOpIndex2BlobMap and the closure using it
The class member mOpIndex2BlobMap is not really used any more, but
it holds blobs which are shared pointers. There is a potential
case of failing to release the blob resource because of the useless
mOpIndex2BlobMap is still holding it. We'd better get rid of such
unnecessary reference before it becomes a real obscure bug later.