Add CI for clang-format checking
This github action checks the coding style in a PR and fails if it
does not meet the coding style in the .clang-format file.
Maintainers should not merge the PR until it is fixed and this
action passes.

Refer to the Coding Style section in the README.md for more details.

Signed-off-by: reagan.lopez@intel.com <reagan.lopez@intel.com>