[TO-BE-REVERTED] Remove transposes if the nodes are intermediate
If the node is a subgraph input convert NHWC to NCHW and
If the node is a subgraph output convert NCHW to NHWC.

This is a hack, as there are corner cases where this
approach breaks. Actual fix is to consistently use the input
layout of the graph as VPUXCompiler can compile the graph
accordingly.

Signed-off-by: Anisha Kulkarni <anisha.dattatraya.kulkarni@intel.com>